Change the PHP open tags to the long version in the fixtures files
The short_open_tag=on is not recommended and most of the linux
distributions ship the php.ini with this option turned off.
This means that the `php -l test/rsrc/index-fail.php` never throws error
because there is no PHP code in that file.
The long open tag (<?php) always works.